requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
psycopg2-binary==2.9.9Werkzeug==3.0.3
WeasyPrint==62.3
Jinja2==3.1.4
Pillow==10.4.0
python-dotenv==1.0.1
